freelance prj/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
apify-client==1.6.3
transformers==4.36.2
torch==2.2.0
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
